initial commit with cypress browser test suite support
- browser tests are located in cypress/integration.  There are two files,
  browser.spec.ts and universal.spec.ts.  Those modules import the tests
  from where the tests lay in the source trees, next to the modules being
  tested.
- we might be able to try setting the cypress test directory to ./src/ and
  passing it a testFiles array to limit the files to just the ones we
  want.  I have not tested that yet.  I was surprised that most of the
  existing PhantomJS tests imported straight in and ran.
- I attempted to follow instructions for configuring code coverage here:
  https://docs.cypress.io/guides/tooling/code-coverage.html#Introduction
  This did not work.  Looks like it needs some more in-depth work to figure
  that out.  We can probably get nyc working for coverage on the node tests
  pretty easily, and then hopefully it won't be too difficult to get for
  browser tests as well.  I'm noticing as I write this, that we already
  have babel-plugin-istanbul installed, so we may be able to use that to
  get us to Constantinople.
- new / updated npm scripts:
  cypress:open - runs cypress tests in interactive mode
  cypress:run - runs cypress tests in headless mode
  test:browser-all - presently runs cypress:run
  test - runs test:module, test:node-all, test:browser-all and the original
  PhantomJS integration tests.
- Original PhantomJS integration tests are not yet ported to Cypress, I
  suspect that they do not work because Cypress does some quite odd things
  when faced with asynchronous code.
- Camera Access tests also do not work in Cypress yet, I believe for same
  reasons -- camera access depends on asynchronous functionality, so some
  porting work for those two test-suites will need to be done.
- moved result_collector.spec.ts to analytics/test/browser (existing tests
  require browser functionality to run, i'm not sure if the code in the
  module itself does or not)
- moved array_helper.spec.ts to common/test
- moved cv_utils.spec.js to common/test
- noted in exif_helper module that it has code that is browser only
- moved exif_helper.spec.ts to browser specific test directory
- moved barcode_locator.spec.ts to locator/test
- run all tests successfully locally
- There is an error message at the end of running PhantomJS tests --
  an IndexSizeError: DOM Exception 1.  That error has been here the whole
  time, and it's not clear what test it is, or why it's failing, and
  especially why it's throwing an error but being treated as success.
  In previous versions, this test was in hte middle of the output, so
  it was hardly visible.  Now, it's at the end of output, for some reason.
  I'm not presently worried about this error, because it's still consistent
  with previously functioning versions.
