+++ .gitignore
@@ -7,4 +7,4 @@ _site/
 .vscode/
 dist/
 lib/quagga.js
-
+cypress/videos
