+++ .npmignore
@@ -1,3 +1,4 @@
+cypress/
 test/
 example/
 examples/
