+++ package.json
@@ -15,6 +15,7 @@
     "@babel/preset-env": "^7.9.6",
     "@babel/preset-typescript": "^7.8.3",
     "@babel/runtime": "^7.9.6",
+    "@cypress/code-coverage": "^3.7.0",
     "@std/esm": "^0.26.0",
     "@types/chai": "^4.2.9",
     "@types/gl-vec2": "^1.3.0",
@@ -26,15 +27,18 @@
     "babel-loader": "^8.0.6",
     "babel-plugin-add-module-exports": "^1.0.2",
     "babel-plugin-istanbul": "^6.0.0",
+    "browserify-istanbul": "^3.0.1",
     "chai": "^4.2.0",
     "core-js": "^3.6.5",
     "cross-env": "^7.0.0",
+    "cypress": "^4.5.0",
     "eslint": "^6.8.0",
     "grunt": "^1.0.4",
     "grunt-cli": "^1.3.2",
     "grunt-contrib-nodeunit": "1.0.0",
     "grunt-karma": "2.0.0",
     "isparta-loader": "^2.0.0",
+    "istanbul-lib-coverage": "^3.0.0",
     "karma": "1.3.0",
     "karma-chai": "0.1.0",
     "karma-chrome-launcher": "2.2.0",
@@ -47,6 +51,7 @@
     "karma-source-map-support": "^1.4.0",
     "karma-webpack": "^4.0.2",
     "mocha": "5.2.0",
+    "nyc": "^15.0.1",
     "phantomjs": "^2.1.7",
     "sinon": "5.1.1",
     "sinon-chai": "^3.5.0",
@@ -62,16 +67,18 @@
   },
   "scripts": {
     "check-types": "tsc",
-    "test:browser-specific": "need something to run a browser suite here",
-    "test:browser-universal": "need something to run a browser suite here",
+    "cypress:open": "cypress open --env BUILD_ENV=production",
+    "cypress:run": "cypress run --env BUILD_ENV=production",
+    "test:browser-specific": "NOT WORKING -- something like npx cypress run --config testFiles=[browser]",
+    "test:browser-universal": "NOT WORKING -- something like npx cypress run --config testFiles=[universal]",
     "test:import": "npx mocha -r @std/esm test/test-import.mjs",
     "test:node-specific": "npx ts-mocha -p test/tsconfig.json src/**/test/node/*.spec.*",
     "test:node-universal": "npx ts-mocha -p test/tsconfig.json -r chai/register-expect src/**/test/*.spec.*",
     "test:require": "npx mocha test/test-require",
-    "test:browser-all": "npm run test:browser-universal && npm run test:browser-specific",
+    "test:browser-all": "npm run cypress:run",
     "test:module": "npm run test:require && npm run test:import",
     "test:node-all": "npm run test:node-universal && npm run test:node-specific",
-    "test": "npm run test:module && npm run test:node-all && npx grunt test",
+    "test": "npm run test:module && npm run test:node-all && npm run test:browser-all && npm run integrationtest",
     "integrationtest": "grunt integrationtest",
     "build:dev": "npx cross-env BUILD_ENV=development webpack --config configs/webpack.config.js",
     "build:prod": "npx cross-env BUILD_ENV=production webpack --config configs/webpack.config.min.js",
